feat(redux): change type of state for favorites store
i change type of state for favorites store to be abble to show directly movies in favorites page.
Before, i save only id of movie in the store but if i want to be abble to show data from this, i
need to request API. Now, i save directly movie inside the store.